refactor(#1704): campsite split — move delta_scan/scan.rs inline tests to two sibling files
Verbatim move (#1116/#1135). scan.rs was 1922 lines (1494 of them one inline
mod tests) against the ~800-line source target; it is now 437, and the #1704
delta seam can be added without growing an over-threshold file.
